More cleanup.

+++ include/lbann/models/model.hpp
@@ -203,7 +203,7 @@ public:
   /** @brief Are background I/O activities enabled by the input layers */
   bool background_io_activity_allowed() { return m_background_io_allowed; }
 
-  int get_num_iterations_per_epoch(execution_mode mode) const;
+  size_t get_num_iterations_per_epoch(execution_mode mode) const;
 
   // ===========================================
   // Setup
